Switched our information in our database to doubles. Additionally there is Player comparison page to compare two player stats.

+++ app/src/main/java/com/example/tennistracker/Player.kt
@@ -2,12 +2,12 @@ package com.example.tennistracker
 
 data class Player(
     val name : String? = null,
-    val pointsWon : Int,
-    val pointsPlayed : Int,
-    val totalFirstServeMissed: Int,
-    val totalSecondServeMissed: Int,
-    val totalReturnMissed: Int,
-    val servesTotal: Int
+    val pointsWon : Double,
+    val pointsPlayed : Double,
+    val totalFirstServeMissed: Double,
+    val totalSecondServeMissed: Double,
+    val totalReturnMissed: Double,
+    val servesTotal: Double
 
 
 )
